scripts/nwea/nwea1.0/requirements.txt
requests
python-dotenv
pandas
tqdm
jupyter
paramiko
sqlalchemy
cx-Oracle